float to int

+++ Assets/Scripts/Player.cs
@@ -2,9 +2,9 @@ using UnityEngine;
 
 public class Player : MonoBehaviour
 {
-    public float health = 100f;
+    public int health = 100;
 
-    public void TakeDamage(float damage)
+    public void TakeDamage(int damage)
     {
         health -= damage;
         Debug.Log("Player Health: " + health);
